first draft of S3 image only tests
