+++ localstack/testing/pytest/snapshot.py
@@ -11,10 +11,12 @@ from _pytest.reports import TestReport
 from _pytest.runner import CallInfo
 from pluggy import Result
 
+from localstack.constants import TEST_AWS_REGION_NAME
 from localstack.testing.snapshots import SnapshotAssertionError, SnapshotSession
 from localstack.testing.snapshots.report import render_report
 from localstack.testing.snapshots.transformer import RegexTransformer
 from localstack.testing.snapshots.transformer_utility import SNAPSHOT_BASIC_TRANSFORMER
+from localstack.utils.bootstrap import is_api_enabled
 
 
 def is_aws():
@@ -99,7 +101,10 @@ def pytest_runtest_call(item: Item) -> None:
 
 @pytest.fixture(name="region", scope="session")
 def fixture_region(aws_client):
-    return aws_client.sts.meta.region_name
+    if is_aws() or is_api_enabled("sts"):
+        return aws_client.sts.meta.region_name
+    else:
+        return TEST_AWS_REGION_NAME
 
 
 @pytest.fixture(scope="function")
