+++ tests/aws/services/s3/conftest.py
@@ -1,0 +1,3 @@
+import os
+
+TEST_S3_IMAGE = os.path.exists("/usr/lib/localstack/.s3-version")
